worker/pyproject.toml
[project]
name = "worker"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "shrpow", email = "shrpow@ya.ru" }
]
dependencies = [
    "pycryptodome>=3.20.0",
    "curl-cffi>=0.7.2",
    "backoff>=2.2.1",
    "pytz>=2024.2",
    "loguru>=0.7.2",
    "types-pytz>=2024.2.0.20240913",
    "pyzmq>=26.2.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "pytest>=8.3.3",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/worker"]
